chore: field switching & hide label
