--- src/blocks/blocks/form/common.ts
@@ -1,28 +1,0 @@
-import { omit } from 'lodash';
-import { createBlock } from '@wordpress/blocks';
-import { dispatch } from '@wordpress/data';
-import { getChoice } from '../../helpers/helper-functions';
-
-export const switchFormFieldTo = ( type?: string, clientId ?:string, attributes?: any ) => {
-
-	if ( ! type || ! clientId || ! attributes ) {
-		return;
-	}
-
-	const { replaceBlock } = dispatch( 'core/block-editor' );
-
-	const blockName = getChoice([
-		[ 'textarea' === type, 'form-textarea' ],
-		[ 'select' === type || 'checkbox' === type || 'radio' === type, 'form-multiple-choice' ],
-		[ 'form-input' ]
-	]);
-
-
-	const newBlock = createBlock( `themeisle-blocks/${ blockName }`,
-		omit({ ...attributes, type: type }, 'form-textarea' === blockName ? [ 'multipleSelection', 'options', 'type' ] : [ 'multipleSelection', 'options' ])
-	);
-
-	replaceBlock( clientId, newBlock );
-};
-
-export default { switchFormFieldTo };
