+++ src/blocks/blocks/form/input/inspector.js
@@ -15,7 +15,7 @@ import {
 	TextControl,
 	ToggleControl
 } from '@wordpress/components';
-import { switchFormFieldTo } from '../common';
+import { HideFieldLabelToggle, switchFormFieldTo } from '../common';
 
 /**
  *
@@ -76,6 +76,10 @@ const Inspector = ({
 						{
 							label: __( 'Textarea', 'otter-blocks' ),
 							value: 'textarea'
+						},
+						{
+							label: __( 'Url', 'otter-blocks' ),
+							value: 'url'
 						}
 					] }
 					onChange={ type => {
@@ -93,6 +97,8 @@ const Inspector = ({
 					value={ attributes.label }
 					onChange={ label => setAttributes({ label }) }
 				/>
+
+				<HideFieldLabelToggle attributes={ attributes } setAttributes={ setAttributes } />
 
 				<TextControl
 					label={ __( 'Placeholder', 'otter-blocks' ) }
